Introduce injectable rule-manager and remove base-rule-manager
